+++ server/deployment/runThis.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+
+# Orden:
+# Read data
+# Installs
+# Database
+# Users
+# Proxy
+# Security
+# Project
+
+
+# File generation of input needed to have the same settings on another server