+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
- - Project's initial commit (14/10/2025)
- - Changelog file (14/10/2025)
+ - Dependency install script for server (14/10/2025)
  - Pull request template file (14/10/2025)
+ - Changelog file (14/10/2025)
+ - Project's initial commit (14/10/2025)